Update AddAction.java, Parser.java to include assertions
The code previously did not have any assertions to validate that
the user input met the requirements after parsing through the
Parser class.

If the Parser class failed to throw an error, it will lead to run time
errors or undefined behaviors as the parsed input does not meet the
expected format.

Hence, the assertion will prevent the remaining code from executing