Add new functionality of creating reminders
The application currently lacks a feature to remind users of upcoming
tasks

Without reminders, users may forget important tasks resulting in
missed deadlines

Hence, the chatbot now accepts a new command "reminder" which will
identify Deadline and Event tasks which are marked not done.

This is to identify Deadline and Event tasks that are upcoming but not
completed yet.